# src/main/java/com/bank/AccountMs/model/Cuenta.java
package com.bank.AccountMs.model;


import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.index.Indexed;
import org.springframework.data.mongodb.core.mapping.Document;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@Document(collection = "cuentas")
public class Cuenta {
    @Id
    private String id;

    @Indexed(unique = true)
    private String numeroCuenta;

    private Double saldo;

    @NotNull(message = "El tipo de cuenta es obligatorio")
    private TipoCuenta tipoCuenta;

    @NotNull(message = "El ID del cliente es obligatorio")
    private String clienteId;
}